updated level 3 with changes per feedback
In the "make organism" module, please change 
-"create creature" to "number of organisms"
-"Food color pref" to "Flower color preference"

-Please remove the "ticks" from the model speed bar
-For "flowers by color" and "how many of each insect type", please add the y-axis label and also add "model time units" to the x-axis
-also, please check "how many of each insect type" since there are some types

Is there a way to keep the garden set up and then run the game again?  

For Scenario #3 game
(1) I think we need to add a "hatch" function so we also increase the number of pollinators
(2) Can you make the default for the beetle to be yellow for pref #1, and orange for pref #2?
(3) Can you make the default for the butterfly to be yellow for pref #1, and blue for pref #2?
(4) Can you change the cost of each flower to $50?